requirements.txt
langchain>=1.2.0
langchain-google-genai>=4.0.0
langchain-community>=0.4.0
langchain-ollama>=1.0.0
langgraph>=1.0.0
chromadb>=1.5.0
pypdf>=6.0.0
python-dotenv>=1.2.0
langchain-text-splitters>=0.3.0
langchain-chroma>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
